Rename module that has configuration page
